Split-off solver code to separate sub-module.
